chore: 파일을 그룹으로 분류 #6
