+++ GyroData/Utility/GraphView/Entities/GraphNumber.swift
@@ -1,0 +1,16 @@
+//
+//  GraphNumber.swift
+//  GyroData
+//
+//  Created by Judy on 2022/12/29.
+//
+
+import Foundation
+
+enum GraphNumber {
+    static let borderWidth = CGFloat(2)
+    static let segmentWidth = CGFloat(3)
+    static let initialPoint = [Double.zero, Double.zero, Double.zero]
+    static let initialRange = -4.0...4.0
+    static let lowerBoundRange = -4.0
+}
