+++ GyroData/Utility/GraphView/GraphView.swift
@@ -86,10 +86,3 @@ class GraphView: UIView {
     }
 }
 
-enum GraphNumber {
-    static let borderWidth = CGFloat(2)
-    static let segmentWidth = CGFloat(3)
-    static let initialPoint = [Double.zero, Double.zero, Double.zero]
-    static let initialRange = -4.0...4.0
-    static let lowerBoundRange = -4.0
-}
